fix(opsi): нормализовать numpy scalar при подсчёте AP

+++ module/os_handler/action_point.py
@@ -205,7 +205,7 @@ class ActionPointHandler(UI, MapEventHandler):
         current = OCR_ACTION_POINT_REMAIN.ocr(self.device.image)
         total = current
         if self.config.OS_ACTION_POINT_BOX_USE:
-            total += np.sum(np.array(box) * tuple(ACTION_POINT_BOX.values()))
+            total += int(np.sum(np.array(box) * tuple(ACTION_POINT_BOX.values())))
         oil = box[0]
 
         LogRes(self.config).Oil = oil
@@ -357,7 +357,6 @@ class ActionPointHandler(UI, MapEventHandler):
         for _ in self.loop(timeout=1):
 
             current, _, total = OCR_ACTION_POINT_BUY_REMAIN.ocr(self.device.image)
-
             # 可能的结果: 0/5, 05
             if total == 0:
                 continue
